Moved test files, wrote readme to aid in use
